chore: added comments and section separators
