+++ core/src/mast/mod.rs
@@ -5,8 +5,8 @@ use miden_crypto::hash::rpo::RpoDigest;
 
 mod node;
 pub use node::{
-    get_span_op_group_count, BasicBlockNode, CallNode, DynNode, ExternalNode, JoinNode, LoopNode,
-    MastNode, OpBatch, OperationOrDecorator, SplitNode, OP_BATCH_SIZE, OP_GROUP_SIZE,
+    BasicBlockNode, CallNode, DynNode, ExternalNode, JoinNode, LoopNode, MastNode, OpBatch,
+    OperationOrDecorator, SplitNode, OP_BATCH_SIZE, OP_GROUP_SIZE,
 };
 use winter_utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable};
 
@@ -18,14 +18,17 @@ mod tests;
 // MAST FOREST
 // ================================================================================================
 
-/// Represents the types of errors that can occur when dealing with MAST forest.
-#[derive(Debug, thiserror::Error)]
-pub enum MastForestError {
-    #[error(
-        "invalid node count: MAST forest exceeds the maximum of {} nodes",
-        MastForest::MAX_NODES
-    )]
-    TooManyNodes,
+/// Represents one or more procedures, represented as a collection of [`MastNode`]s.
+///
+/// A [`MastForest`] does not have an entrypoint, and hence is not executable. A [`crate::Program`]
+/// can be built from a [`MastForest`] to specify an entrypoint.
+#[derive(Clone, Debug, Default, PartialEq, Eq)]
+pub struct MastForest {
+    /// All of the nodes local to the trees comprising the MAST forest.
+    nodes: Vec<MastNode>,
+
+    /// Roots of procedures defined within this MAST forest.
+    roots: Vec<MastNodeId>,
 }
 
 // ------------------------------------------------------------------------------------------------
@@ -174,15 +177,12 @@ impl Deserializable for MastNodeId {
 // MAST FOREST ERROR
 // ================================================================================================
 
-/// Represents one or more procedures, represented as a collection of [`MastNode`]s.
-///
-/// A [`MastForest`] does not have an entrypoint, and hence is not executable. A [`crate::Program`]
-/// can be built from a [`MastForest`] to specify an entrypoint.
-#[derive(Clone, Debug, Default, PartialEq, Eq)]
-pub struct MastForest {
-    /// All of the nodes local to the trees comprising the MAST forest.
-    nodes: Vec<MastNode>,
-
-    /// Roots of procedures defined within this MAST forest.
-    roots: Vec<MastNodeId>,
+/// Represents the types of errors that can occur when dealing with MAST forest.
+#[derive(Debug, thiserror::Error)]
+pub enum MastForestError {
+    #[error(
+        "invalid node count: MAST forest exceeds the maximum of {} nodes",
+        MastForest::MAX_NODES
+    )]
+    TooManyNodes,
 }
