+++ core/src/mast/node/loop_node.rs
@@ -12,6 +12,12 @@ use crate::{
 // LOOP NODE
 // ================================================================================================
 
+/// A Loop node defines condition-controlled iterative execution. When the VM encounters a Loop
+/// node, it will keep executing the body of the loop as long as the top of the stack is `1``.
+///
+/// The loop is exited when at the end of executing the loop body the top of the stack is `0``.
+/// If the top of the stack is neither `0` nor `1` when the condition is checked, the execution
+/// fails.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct LoopNode {
     body: MastNodeId,
@@ -38,19 +44,19 @@ impl LoopNode {
 }
 
 impl LoopNode {
-    pub fn body(&self) -> MastNodeId {
-        self.body
-    }
-
+    /// Returns a commitment to this Loop node.
+    ///
+    /// The commitment is computed as a hash of the loop body and an empty word ([ZERO; 4]) in
+    /// the domain defined by [Self::DOMAIN] - i..e,:
+    ///
+    /// hasher::merge_in_domain(&[on_true_digest, Digest::default()], LoopNode::DOMAIN)
     pub fn digest(&self) -> RpoDigest {
         self.digest
     }
 
-    pub fn to_display<'a>(&'a self, mast_forest: &'a MastForest) -> impl fmt::Display + 'a {
-        LoopNodePrettyPrint {
-            loop_node: self,
-            mast_forest,
-        }
+    /// Returns the ID of the node presenting the body of the loop.
+    pub fn body(&self) -> MastNodeId {
+        self.body
     }
 }
 
@@ -58,6 +64,13 @@ impl LoopNode {
 // ================================================================================================
 
 impl LoopNode {
+    pub(super) fn to_display<'a>(&'a self, mast_forest: &'a MastForest) -> impl fmt::Display + 'a {
+        LoopNodePrettyPrint {
+            loop_node: self,
+            mast_forest,
+        }
+    }
+
     pub(super) fn to_pretty_print<'a>(
         &'a self,
         mast_forest: &'a MastForest,
